+++ NEWS.md
@@ -10,6 +10,7 @@
 
 -   Added `summary`, `vcov` and `coef` method for `familiarModel` objects. These respectively apply `summary`, `vcov` and `coef` to the stored model.
 -   Added relative absolute error, relative squared error and root relative squared error as performance metrics.
+-   Models to predict the score of hyperparameter sets are now object-oriented. This change is not visible to the user.
 
 ## Bug fixes